Add files via upload
Change to the way the script works.
It now holds a window open while checking ED log files.

If the log file is quiet (unchanging), the script reports that it is checking the files.  Note, this implies not much is happening in the game.

If there are a lot of events in the file, then that *should* be reflected in changes to the device LEDs.